feat(RSS): Rework RSS UI
