+++ src/components/RSS/Mobile.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <h1>RSS Mobile</h1>
+</template>
+
+<style scoped>
+
+</style>
