+++ src/types/vuetorrent/RssArticle.ts
@@ -1,6 +1,8 @@
 import { FeedArticle } from '@/types/qbit/models'
 
 export interface RssArticle extends FeedArticle {
+  /** UID of the parent feed */
+  feedId: string
   /** Article publication date parsed by dayjs */
   parsedDate: Date
 }
